requirements.txt
# LogWatchdog - Core Dependencies
# Windows-Only Log Monitoring Solution

# Core monitoring and notification
plyer>=2.0.0                    # Cross-platform notifications
python-dotenv>=0.19.0           # Environment variable management
rich>=10.0.0                    # Rich console output and formatting

# Windows system monitoring dependencies
psutil>=5.8.0                   # Windows system and process monitoring

# File monitoring and processing
watchdog>=2.1.0                 # File system monitoring
colorama>=0.4.4                 # Windows console color support

# Email and communication
# Note: smtplib is built-in to Python, no external package needed
email-validator>=1.1.3          # Email address validation

# Configuration and data handling
# Note: configparser is built-in to Python, no external package needed

# Logging and debugging
# Note: logging is built-in to Python, no external package needed

# Optional dependencies (uncomment as needed)
# pyyaml>=6.0                     # YAML configuration support
# jsonschema>=4.0.0               # JSON schema validation
# structlog>=21.1.0                # Structured logging
# loguru>=0.5.3                   # Advanced logging features

# Development and testing (optional)
# Uncomment for development environment
# pytest>=6.0.0                 # Testing framework
# pytest-cov>=2.10.0            # Coverage reporting
# black>=21.0.0                 # Code formatting
# flake8>=3.8.0                 # Linting
# mypy>=0.800                   # Type checking

# Build and distribution (optional)
# Uncomment for building executables
# pyinstaller>=5.0.0            # Create standalone executables
# cx_Freeze>=6.0                # Alternative executable creation
